requirements.txt
alabaster==0.7.12
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudpickle==2.0.0
Cython==0.29.24
decorator==5.1.0
docopt==0.6.2
docutils==0.17.1
et-xmlfile==1.1.0
fasteners==0.16.3
glfw==2.4.0
gym==0.21.0
idna==3.3
ImageHash==4.2.1
imageio==2.10.4
imagesize==1.3.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.29.0
jedi==0.18.0
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mujoco-py==2.0.2.13
numpy==1.21.4
numpydoc==1.1.0
openpyxl==3.0.9
packaging==21.2
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.11
pluggy==1.0.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-instafail==0.3.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
requests==2.26.0
scipy==1.7.2
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.3.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
torch==1.10.0
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
yarg==0.1.9